[api] Move DNS configuration to dns.go
